Added Spikes
Now there are enemies to avoid. Cool! Also had to put in a try/catch block for the enemies, because one map does not have it. Did not find a way to get a count of the objects in a layer, and if there were no objects then it caused a crash.